src/TicketMonarch/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
pandas==2.1.4
mysql-connector-python==9.0.0
python-dotenv==1.0.1

